THE FIRST OF MANY CHANGES IN THE UXREWORK
Started the banner section and got the text and logo in. Circle is part of
bg image. Maybe try to make in a div but not right now.
Text wrapping is done with the use of p spans. Cool method. Working on red
lines next. Then focus on the new imageBoard stuff. Design javascript in
notebook.