HEADER ALMOST COMPLETE
Header almost complete. Need to decide on hover color for links in nav.
Used a padder li to space nav links across logo. It works for now.
Need to work on image board still. Got nothing done for it. Add div and
start to write algorithm for js functionality. Want to hover preview the
size of the box then on click show full size image with caption /
description.